finished ui tests for product screen

+++ app/src/main/java/com/gmail/bodziowaty6978/fitnessappv2/feature_diary/domain/model/Product.kt
@@ -10,7 +10,7 @@ data class Product(
     @ColumnInfo(name = "author") val author: String = "Joe Mama",
     @ColumnInfo(name = "name") val name: String = "White rice",
     @ColumnInfo(name = "brand") val brand: String? = null,
-    @ColumnInfo(name = "containerWeight") val containerWeight: Double = 0.0,
+    @ColumnInfo(name = "containerWeight") val containerWeight: Int = 0,
     @ColumnInfo(name = "position") val position: Int = 0,
     @ColumnInfo(name = "unit") val unit: String = "g",
     @ColumnInfo(name = "nutritionValues") val nutritionValues: NutritionValues = NutritionValues(),
